requirements.txt
APScheduler==3.6.3
attrs==19.3.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2019.9.11
chardet==3.0.4
idna==2.8
importlib-metadata==1.5.0
more-itertools==8.2.0
packaging==20.1
pkg_resources==0.0.0
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
PySocks==1.7.1
pytest==5.3.5
python-dotenv==0.12.0
python-telegram-bot==13.7
pytz==2019.3
requests==2.22.0
six==1.14.0
tornado==6.1
twilio==6.35.5
tzlocal==3.0
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
